% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article.R
\name{get_articles}
\alias{get_articles}
\title{Get news titles on a specific company from Google News}
\usage{
get_articles(company_name, n = 5)
}
\arguments{
\item{company_name}{name of the company you'd like to get news titles}

\item{n}{numbers of news titles to return, default is 5}
}
\description{
Use this function to have a flash understanding on a certain company event. Its purpose it's to use it after having detected companies had a negative performance, and get some information about them.
}
