% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R
\name{GetInfoPrices}
\alias{GetInfoPrices}
\title{Function to download stocks closing price}
\usage{
GetInfoPrices(tck, tm)
}
\arguments{
\item{tck}{list of companies tickers}

\item{tm}{number of days of the range}
}
\description{
This function is designed to download the last closing prices and a selected closing prices in the past of a list of companies.
}
