% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{GetPerformance}
\alias{GetPerformance}
\title{Function to get performance and filter the data based on that.}
\usage{
GetPerformance(data, percentage)
}
\arguments{
\item{data}{dataframe resulting from the get_price function.}

\item{percentage}{treshold to filter stocks. The function return all the stocks that perform less than the given percentage.}
}
\description{
The purpose of this function is to filter stocks based on performance. Given a certain percentage, the function return all the stocks that performed less than that percentage.
}
