% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stocks.R
\name{PlotCompanies}
\alias{PlotCompanies}
\title{Plot stocks that require more interest}
\usage{
PlotCompanies(tck, years)
}
\arguments{
\item{tck}{list of the stocks tickers}

\item{years}{number of years to visualize}
}
\description{
This function return plot for each companies selected. It's helpful to get a larger view of the stocks trend.
}
